% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{add_features}
\alias{add_features}
\title{Adds extra features}
\usage{
add_features(data)
}
\arguments{
\item{data}{}
}
\value{
Data frame with additional features.
}
\description{
Minimum, maximum and mean temperatures are calculated across all weather
stations. Trend is a numeric variable increasing with date.
}
