% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{add_lags}
\alias{add_lags}
\title{Add lags}
\usage{
add_lags(data, lags = NULL)
}
\arguments{
\item{data}{(data frame) Data with variables to create lags for.}

\item{lags}{(list) List of lags to include for each variable variables.
Element name should be variable and numeric values will give lags. Lags
should be specified as number of half-hourly periods, e.g., c(1, 3, 5) will
create 30, 90 and 150 minute lags.}
}
\value{
Data frame with additional lags.
}
\description{
Add lags
}
