% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{cv_ts_folds}
\alias{cv_ts_folds}
\title{Get cross validation folds for time series}
\usage{
cv_ts_folds(x, start_date, horizon, iterations)
}
\arguments{
\item{x}{(datetime) Vector of dates to create indexes for. Must be ordered.}

\item{start_date}{(date) Date to start creating folds from.}

\item{horizon}{(int) Number of days making up each out-of-sample test set.}

\item{iterations}{(int) Number of training and test sets to create.}
}
\value{
List with train and test elements for each iteration.
}
\description{
Get cross validation folds for time series
}
\examples{
library(lubridate)
dts <- ymd("2020-01-01") + minutes(30)*0:(48*14-1)
cv_folds <- cv_ts_folds(dts, ymd("2020-01-07"), 2, 3)
}
