% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule-battery.R
\name{format_charge_data}
\alias{format_charge_data}
\title{Format battery charge data}
\usage{
format_charge_data(B)
}
\arguments{
\item{B}{(matrix) Matrix of half-hourly battery charges/discharges.}
}
\value{
Data frame of battery charges/discharges.
}
\description{
Formats battery charge data to required competition format.
}
