% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hh_to_period}
\alias{hh_to_period}
\title{Half-hour to period of day}
\usage{
hh_to_period(x)
}
\arguments{
\item{x}{vector of datetimes.}
}
\value{
vector giving period of day.
}
\description{
Half-hour to period of day
}
