% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{load_data}
\alias{load_data}
\title{Load data}
\usage{
load_data(path, locations = 1:6, inc_pv_cond = F, lags = NULL)
}
\arguments{
\item{path}{(chr) Path to data.}

\item{locations}{(int) Integers specifying which weather station locations to
include. Possible values are 1-6.}

\item{inc_pv_cond}{(logical) If true, irradiance at PV facility
(irradiance_wm2) and panel temperature (panel_temp_c) are also returned.}

\item{lags}{(list) List of lags to include for each variable variables.
Element name should be variable and numeric values will give lags. Lags
should be specified as number of half-hourly periods, e.g., c(1, 3, 5) will
create 30, 90 and 150 minute lags.}
}
\value{
Data frame.
}
\description{
Loads demand, PV and weather data.
}
