% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{load_demand_data}
\alias{load_demand_data}
\title{Load demand data}
\usage{
load_demand_data()
}
\value{
Data frame containing demand (MW) data and required features for
  training demand model.
}
\description{
Load demand data
}
