% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{load_pv_data}
\alias{load_pv_data}
\title{Load PV data}
\usage{
load_pv_data()
}
\value{
Data frame containing PV generation (MW) data and required features
  for training PV generation model.
}
\description{
Load PV data
}
