% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pod.R
\docType{data}
\name{pod}
\alias{pod}
\title{Data for POD battery scheduling competition}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 37536 rows and 16 columns.
}
\usage{
pod
}
\description{
A dataset containing the electricity demand, PV power, weather and public
holiday data. Data collected from 6 weather stations are included.
}
\keyword{datasets}
