% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{pred_demand_period}
\alias{pred_demand_period}
\title{Predict demand with models for each period}
\usage{
pred_demand_period(data, train_idx, test_idx, ...)
}
\arguments{
\item{data}{(data frame) Data frame containing predictors and response variable `demand_mw`. Must also contain `period` column.}

\item{train_idx}{(integer) Vector of integers specifying which rows to use for training model}

\item{test_idx}{(integer) Vector of integers specifying which rows to use for testing model}

\item{...}{Additional arguments passed to `lgb.fit`.}
}
\value{

}
\description{
Predict demand with models for each period
}
