% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{pred_point}
\alias{pred_point}
\title{Predict point prediction using lightgbm}
\usage{
pred_point(data, train_idx, test_idx, response_idx, ...)
}
\arguments{
\item{data}{}

\item{train_idx}{}

\item{test_idx}{}

\item{response_idx}{}

\item{...}{}
}
\description{
Predict point prediction using lightgbm
}
