% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{pred_pv}
\alias{pred_pv}
\title{Predict PV power}
\usage{
pred_pv(data, train_idx, test_idx, ...)
}
\arguments{
\item{data}{(data frame) Data frame containing predictors and response variable `pv_power_mw`.}

\item{train_idx}{(integer) Vector of integers specifying which rows to use for training model}

\item{test_idx}{(integer) Vector of integers specifying which rows to use for testing model}

\item{...}{Additional arguments passed to `lgb.fit`.}
}
\value{

}
\description{
Predict PV power
}
