% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{pred_pv_quantile}
\alias{pred_pv_quantile}
\title{Predict PV power with quantile selection}
\usage{
pred_pv_quantile(data, train_idx, test_idx, alpha = seq(0.5, 0.9, 0.01), ...)
}
\arguments{
\item{data}{(data frame) Data frame containing predictors, response variable `pv_power_mw` and `datetime`.}

\item{train_idx}{(integer) Vector of integers specifying which rows to use for training model}

\item{test_idx}{(integer) Vector of integers specifying which rows to use for testing model}

\item{alpha}{(numeric) Vector of quantiles 0.01-0.99.}

\item{...}{Additional arguments passed to `lgb.fit`.}
}
\value{

}
\description{
Predict PV power with quantile selection
}
