% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{pred_quantile}
\alias{pred_quantile}
\title{Predict quantile predictions using lightgbm}
\usage{
pred_quantile(data, train_idx, test_idx, response_idx, alpha, ...)
}
\arguments{
\item{data}{(data frame) Data frame containing predictors, response variable `pv_power_mw` and `datetime`.}

\item{train_idx}{(integer) Vector of integers specifying which rows to use for training model}

\item{test_idx}{(integer) Vector of integers specifying which rows to use for testing model}

\item{response_idx}{(integer) Column index of response variable in `data`.}

\item{alpha}{(numeric) Vector of quantiles 0.01-0.99.}

\item{...}{Additional arguments passed to `lgb.fit`.}
}
\description{
Predict quantile predictions using lightgbm
}
