% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{predict_lgbm}
\alias{predict_lgbm}
\title{Predict using lightgbm}
\usage{
predict_lgbm(data, train_idx, test_idx, response_idx, ...)
}
\arguments{
\item{data}{}

\item{train_idx}{}

\item{test_idx}{}

\item{response_idx}{}

\item{...}{}
}
\value{

}
\description{
Predict using lightgbm
}
