% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{predict_quantile}
\alias{predict_quantile}
\title{Predict quantile predictions using lightgbm}
\usage{
predict_quantile(data, train_idx, test_idx, response_idx, alpha, ...)
}
\arguments{
\item{data}{}

\item{train_idx}{}

\item{test_idx}{}

\item{response_idx}{}

\item{...}{}
}
\description{
Predict quantile predictions using lightgbm
}
