% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule-battery.R
\name{schedule_battery}
\alias{schedule_battery}
\title{Schedule battery}
\usage{
schedule_battery(data)
}
\arguments{
\item{data}{(data frame) Data frame containing the columns datetime, period,
pv_power_mw}
}
\value{
List of matrices showing the battery charges/discharges, cumulative
  charge, loads and PV generation.
}
\description{
Function that schedules the charging and discharging of a battery at
half-hourly intervals to reduce peak demand.
}
\details{
The operation of the battery adheres to constraints outlined in the POD data
science challenge (2021). Each day of charging and discharging is managed
separately with the battery starting from zero charge each day. Any changes
in charge must be between -2.5 MW and 2.5 MW. Between periods 1-41 only
charging can occur. Discharging can only occur between periods 32-42.

To charge the battery, the algorithm loops through each periods between 2 and
41. It takes a proportion of solar PV produced during each half-hourly period
such that the total energy sums to the battery capacity. If the total PV
energy is less than 6 MWh then additional energy is drawn from the grid to
ensure the battery is fully charged.
}
