% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_fcst_file}
\alias{score_fcst_file}
\title{Score forecasts}
\usage{
score_fcst_file(file)
}
\arguments{
\item{fcst}{(chr) File containing forecasts in competition format.}

\item{task}{(int) Forecast task number (0-5).}
}
\value{
Named vector with scores for each date.
}
\description{
Scores forecasts.
}
