% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{yday_ly_adj}
\alias{yday_ly_adj}
\title{Adjust yday for leap years}
\usage{
yday_ly_adj(x)
}
\arguments{
\item{x}{(datetime) vector of datetime values.}
}
\description{
Adjusts yday values for leap years. 29 February is now assigned 59.5 and following dates are assigned their original yday minus 1. This ensures yday values are consistent with dates across all years.
}
