% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{load_data}
\alias{load_data}
\title{Load data}
\usage{
load_data(path, locations = 1:6, inc_pv_cond = F)
}
\arguments{
\item{path}{(chr) Path to data.}

\item{locations}{(int) Integers specifying which weather station locations to
include. Possible values are 1-6.}

\item{inc_pv_cond}{(logical) If true, irradiance at PV facility
(irradiance_wm2) and panel temperature (panel_temp_c) are also returned.}
}
\value{
Data frame containing all joined data.
}
\description{
Loads demand, PV and weather data.
}
